fix: refine validation rules as part of pre-1.0 work
This commit covers the following rules:
- array-boundary
- array-items
- property-attributes
- string-boundary
- accept-parameter
- authorization-parameter
- content-type-parameter
