+++ packages/ruleset/src/functions/property-attributes.js
@@ -4,12 +4,18 @@ const {
   isIntegerSchema,
   isFloatSchema,
   isDoubleSchema,
-  isArraySchema,
   isObjectSchema
 } = require('@ibm-cloud/openapi-ruleset-utilities');
+const { LoggerFactory } = require('../utils');
 
-module.exports = function(schema, _opts, { path }) {
-  return validateSubschemas(schema, path, checkPropertyAttributes);
+let ruleId;
+let logger;
+module.exports = function(schema, _opts, context) {
+  if (!logger) {
+    ruleId = context.rule.name;
+    logger = LoggerFactory.newInstance().getLogger(ruleId);
+  }
+  return validateSubschemas(schema, context.path, checkPropertyAttributes);
 };
 
 /**
@@ -18,18 +24,23 @@ module.exports = function(schema, _opts, { path }) {
  * 1) minimum/maximum should not be defined for a non-numeric (number, integer) schema
  * 2) minimum <= maximum
  * 3) minItems/maxItems should not be defined for a non-array schema
- * 4) minItems <= maxItems
- * 5) minProperties/maxProperties should not be defined for a non-object schema
- * 6) minProperties <= maxProperties
+ * 4) minProperties/maxProperties should not be defined for a non-object schema
+ * 5) minProperties <= maxProperties
  *
  * @param {*} schema the schema to check
  * @param {*} path the array of path segments indicating the "location" of the schema within the API definition
  * @returns an array containing the violations found or [] if no violations
  */
 function checkPropertyAttributes(schema, path) {
+  logger.debug(
+    `${ruleId}: checking attributes of schema at location: ${path.join('.')}`
+  );
+
   const errors = [];
 
   if (isNumericSchema(schema)) {
+    logger.debug('schema is numeric');
+
     // 2) minimum <= maximum
     if (schema.minimum && schema.maximum && schema.minimum > schema.maximum) {
       errors.push({
@@ -53,36 +64,10 @@ function checkPropertyAttributes(schema, path) {
     }
   }
 
-  if (isArraySchema(schema)) {
-    // 4) minItems <= maxItems
-    if (
-      schema.minItems &&
-      schema.maxItems &&
-      schema.minItems > schema.maxItems
-    ) {
-      errors.push({
-        message: 'minItems cannot be greater than maxItems',
-        path: [...path, 'minItems']
-      });
-    }
-  } else {
-    // 3) minItems/maxItems should not be defined for a non-array schema
-    if (schema.minItems) {
-      errors.push({
-        message: 'minItems should not be defined for a non-array schema',
-        path: [...path, 'minItems']
-      });
-    }
-    if (schema.maxItems) {
-      errors.push({
-        message: 'maxItems should not be defined for a non-array schema',
-        path: [...path, 'maxItems']
-      });
-    }
-  }
-
   if (isObjectSchema(schema)) {
-    // 6) minProperties <= maxProperties
+    logger.debug('schema is an object schema');
+
+    // 5) minProperties <= maxProperties
     if (
       schema.minProperties &&
       schema.maxProperties &&
@@ -94,7 +79,7 @@ function checkPropertyAttributes(schema, path) {
       });
     }
   } else {
-    // 5) minProperties/maxProperties should not be defined for a non-object schema
+    // 4) minProperties/maxProperties should not be defined for a non-object schema
     if (schema.minProperties) {
       errors.push({
         message: 'minProperties should not be defined for a non-object schema',
