+++ packages/ruleset/src/rules/authorization-parameter.js
@@ -1,14 +1,14 @@
 const {
   parameters
 } = require('@ibm-cloud/openapi-ruleset-utilities/src/collections');
-const { oas2, oas3 } = require('@stoplight/spectral-formats');
+const { oas3 } = require('@stoplight/spectral-formats');
 const { disallowedHeaderParameter } = require('../functions');
 
 module.exports = {
   description:
     "Operations should not explicitly define the 'Authorization' header parameter",
   message: '{{description}}',
-  formats: [oas2, oas3],
+  formats: [oas3],
   given: parameters,
   severity: 'warn',
   then: {
