+++ packages/ruleset/test/array-attributes.test.js
@@ -1,11 +1,12 @@
-const { arrayBoundary } = require('../src/rules');
+const { arrayAttributes } = require('../src/rules');
 const { makeCopy, rootDocument, testRule, severityCodes } = require('./utils');
 
-const rule = arrayBoundary;
-const ruleId = 'ibm-array-boundary';
+const rule = arrayAttributes;
+const ruleId = 'ibm-array-attributes';
 const expectedSeverity = severityCodes.warning;
 const expectedMsgMin = 'Array schemas should define a numeric minItems field';
 const expectedMsgMax = 'Array schemas should define a numeric maxItems field';
+const expectedMsgItems = 'Array schemas must specify the items field';
 
 describe('Spectral rule: array-boundary', () => {
   describe('Should not yield errors', () => {
@@ -85,10 +86,42 @@ describe('Spectral rule: array-boundary', () => {
       const results = await testRule(ruleId, rule, testDocument);
       expect(results).toHaveLength(0);
     });
+
+    it('minItems and maxItems defined', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.Car.properties['wheel_count'] = {
+        type: 'array',
+        minItems: 3,
+        maxItems: 4,
+        items: {
+          type: 'integer'
+        }
+      };
+
+      const results = await testRule(ruleId, rule, rootDocument);
+      expect(results).toHaveLength(0);
+    });
+
+    it('minItems <= maxItems', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.Car.properties['wheel_count'] = {
+        type: 'array',
+        minItems: 3,
+        maxItems: 3,
+        items: {
+          type: 'integer'
+        }
+      };
+
+      const results = await testRule(ruleId, rule, rootDocument);
+      expect(results).toHaveLength(0);
+    });
   });
 
   describe('Should yield errors', () => {
-    it('Response schema array property with no min/maxItems', async () => {
+    it('Array property with no min/maxItems', async () => {
       const testDocument = makeCopy(rootDocument);
 
       // Make a copy of Movie named Movie2, and make Movie2 the response schema
@@ -385,6 +418,257 @@ describe('Spectral rule: array-boundary', () => {
       expect(results[0].path.join('.')).toBe(
         'paths./v1/drinks.post.responses.201.content.application/json.schema.properties.main_prop'
       );
+    });
+    it('allOf element without items property', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      delete testDocument.components.schemas.DrinkCollection.allOf[1].properties
+        .drinks.items;
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(1);
+      expect(results[0].code).toBe(ruleId);
+      expect(results[0].message).toBe(expectedMsgItems);
+      expect(results[0].severity).toBe(expectedSeverity);
+      expect(results[0].path).toStrictEqual([
+        'paths',
+        '/v1/drinks',
+        'get',
+        'responses',
+        '200',
+        'content',
+        'application/json',
+        'schema',
+        'allOf',
+        '1',
+        'properties',
+        'drinks'
+      ]);
+    });
+    it('Response schema without items property', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.MovieCollection = {
+        type: 'array'
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(1);
+      expect(results[0].code).toBe(ruleId);
+      expect(results[0].message).toBe(expectedMsgItems);
+      expect(results[0].severity).toBe(expectedSeverity);
+      expect(results[0].path).toStrictEqual([
+        'paths',
+        '/v1/movies',
+        'get',
+        'responses',
+        '200',
+        'content',
+        'application/json',
+        'schema'
+      ]);
+    });
+    it('Request schema without items property', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.paths['/v1/movies'].post.requestBody.content[
+        'application/json'
+      ].schema = {
+        type: 'array'
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(1);
+      expect(results[0].code).toBe(ruleId);
+      expect(results[0].message).toBe(expectedMsgItems);
+      expect(results[0].severity).toBe(expectedSeverity);
+      expect(results[0].path).toStrictEqual([
+        'paths',
+        '/v1/movies',
+        'post',
+        'requestBody',
+        'content',
+        'application/json',
+        'schema'
+      ]);
+    });
+    it('Request schema with non-object items property', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.paths['/v1/movies'].post.requestBody.content[
+        'application/json'
+      ].schema = {
+        type: 'array',
+        items: 'not a schema!'
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(1);
+      expect(results[0].code).toBe(ruleId);
+      expect(results[0].message).toBe(expectedMsgItems);
+      expect(results[0].severity).toBe(expectedSeverity);
+      expect(results[0].path).toStrictEqual([
+        'paths',
+        '/v1/movies',
+        'post',
+        'requestBody',
+        'content',
+        'application/json',
+        'schema'
+      ]);
+    });
+    it('additionalProperties schema without items property', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.Movie.additionalProperties = {
+        type: 'array'
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(5);
+      for (const result of results) {
+        expect(result.code).toBe(ruleId);
+        expect(result.message).toBe(expectedMsgItems);
+        expect(result.severity).toBe(expectedSeverity);
+      }
+      expect(results[0].path).toStrictEqual([
+        'paths',
+        '/v1/movies',
+        'post',
+        'requestBody',
+        'content',
+        'application/json',
+        'schema',
+        'additionalProperties'
+      ]);
+      expect(results[1].path).toStrictEqual([
+        'paths',
+        '/v1/movies',
+        'post',
+        'responses',
+        '201',
+        'content',
+        'application/json',
+        'schema',
+        'additionalProperties'
+      ]);
+      expect(results[2].path).toStrictEqual([
+        'paths',
+        '/v1/movies',
+        'get',
+        'responses',
+        '200',
+        'content',
+        'application/json',
+        'schema',
+        'allOf',
+        '1',
+        'properties',
+        'movies',
+        'items',
+        'additionalProperties'
+      ]);
+      expect(results[3].path).toStrictEqual([
+        'paths',
+        '/v1/movies/{movie_id}',
+        'get',
+        'responses',
+        '200',
+        'content',
+        'application/json',
+        'schema',
+        'additionalProperties'
+      ]);
+      expect(results[4].path).toStrictEqual([
+        'paths',
+        '/v1/movies/{movie_id}',
+        'put',
+        'requestBody',
+        'content',
+        'application/json',
+        'schema',
+        'additionalProperties'
+      ]);
+    });
+    it('minItems > maxItems', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.Car.properties['wheel_count'] = {
+        type: 'array',
+
+        items: {
+          type: 'integer'
+        },
+        minItems: 5,
+        maxItems: 4
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(4);
+      const expectedPaths = [
+        'paths./v1/cars.post.requestBody.content.application/json.schema.properties.wheel_count',
+        'paths./v1/cars.post.responses.201.content.application/json.schema.properties.wheel_count',
+        'paths./v1/cars/{car_id}.get.responses.200.content.application/json.schema.properties.wheel_count',
+        'paths./v1/cars/{car_id}.patch.responses.200.content.application/json.schema.properties.wheel_count'
+      ];
+      for (let i = 0; i < results.length; i++) {
+        expect(results[i].code).toBe(ruleId);
+        expect(results[i].message).toBe(
+          'minItems cannot be greater than maxItems'
+        );
+        expect(results[i].severity).toBe(expectedSeverity);
+        expect(results[i].path.join('.')).toBe(expectedPaths[i]);
+      }
+    });
+    it('minItems defined for non-array schema', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.Car.properties['wheel_count'] = {
+        type: 'object',
+        minItems: 3
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(4);
+      const expectedPaths = [
+        'paths./v1/cars.post.requestBody.content.application/json.schema.properties.wheel_count.minItems',
+        'paths./v1/cars.post.responses.201.content.application/json.schema.properties.wheel_count.minItems',
+        'paths./v1/cars/{car_id}.get.responses.200.content.application/json.schema.properties.wheel_count.minItems',
+        'paths./v1/cars/{car_id}.patch.responses.200.content.application/json.schema.properties.wheel_count.minItems'
+      ];
+      for (let i = 0; i < results.length; i++) {
+        expect(results[i].code).toBe(ruleId);
+        expect(results[i].message).toBe(
+          'minItems should not be defined for a non-array schema'
+        );
+        expect(results[i].severity).toBe(expectedSeverity);
+        expect(results[i].path.join('.')).toBe(expectedPaths[i]);
+      }
+    });
+    it('maxItems defined for non-array schema', async () => {
+      const testDocument = makeCopy(rootDocument);
+
+      testDocument.components.schemas.Car.properties['wheel_count'] = {
+        type: 'integer',
+        maxItems: 3
+      };
+
+      const results = await testRule(ruleId, rule, testDocument);
+      expect(results).toHaveLength(4);
+      const expectedPaths = [
+        'paths./v1/cars.post.requestBody.content.application/json.schema.properties.wheel_count.maxItems',
+        'paths./v1/cars.post.responses.201.content.application/json.schema.properties.wheel_count.maxItems',
+        'paths./v1/cars/{car_id}.get.responses.200.content.application/json.schema.properties.wheel_count.maxItems',
+        'paths./v1/cars/{car_id}.patch.responses.200.content.application/json.schema.properties.wheel_count.maxItems'
+      ];
+      for (let i = 0; i < results.length; i++) {
+        expect(results[i].code).toBe(ruleId);
+        expect(results[i].message).toBe(
+          'maxItems should not be defined for a non-array schema'
+        );
+        expect(results[i].severity).toBe(expectedSeverity);
+        expect(results[i].path.join('.')).toBe(expectedPaths[i]);
+      }
     });
   });
 });
